+++ tests/integration/api/test_view_sample.py
@@ -6,12 +6,8 @@ from typing import Any
 
 import pandas as pd
 import pytest
-import pytest_asyncio
 from pandas.testing import assert_series_equal
 from pydantic.error_wrappers import ValidationError
-from sqlglot import parse_one
-
-from featurebyte.query_graph.sql.common import sql_to_string
 
 
 def maybe_tz_convert(ts: Any):
@@ -269,41 +265,6 @@ def test_dimension_view_sample_with_date_range(dimension_table):
         assert "timestamp_column must be specified." in str(exc)
 
 
-@pytest_asyncio.fixture(name="source_table_with_invalid_dates")
-async def source_table_with_invalid_dates_fixture(session, feature_store, catalog):
-    """
-    Fixture for a source table with invalid dates
-    """
-    _ = catalog
-    query = sql_to_string(
-        parse_one(
-            """
-            CREATE TABLE TABLE_INVALID_DATES AS
-            SELECT 1 AS "id", CAST('2021-01-01 10:00:00' AS TIMESTAMP) AS "date_col"
-            UNION ALL
-            SELECT 2 AS "id", CAST('0019-01-01 10:00:00' AS TIMESTAMP) AS "date_col"
-            UNION ALL
-            SELECT 3 AS "id", CAST('0019-01-01 10:00:00' AS TIMESTAMP) AS "date_col"
-            UNION ALL
-            SELECT 4 AS "id", CAST('0019-01-01 10:00:00' AS TIMESTAMP) AS "date_col"
-            UNION ALL
-            SELECT 5 AS "id", CAST('9019-01-01 10:00:00' AS TIMESTAMP) AS "date_col"
-            UNION ALL
-            SELECT 6 AS "id", CAST('2023-01-01 10:00:00' AS TIMESTAMP) AS "date_col"
-            """,
-            read="snowflake",
-        ),
-        source_type=session.source_type,
-    )
-    await session.execute_query(query)
-    ds = feature_store.get_data_source()
-    return ds.get_source_table(
-        table_name="TABLE_INVALID_DATES",
-        database_name=session.database_name,
-        schema_name=session.schema_name,
-    )
-
-
 @pytest.mark.asyncio
 async def test_sample_invalid_dates(session, source_table_with_invalid_dates):
     """
@@ -327,28 +288,3 @@ async def test_sample_invalid_dates(session, source_table_with_invalid_dates):
     )
     sample_df = sample_df.sort_values("id").reset_index(drop=True)
     pd.testing.assert_frame_equal(sample_df, expected)
-
-
-@pytest.mark.asyncio
-async def test_preview_invalid_dates(session, source_table_with_invalid_dates):
-    """
-    Test preview with invalid dates
-    """
-    preview_df = source_table_with_invalid_dates.preview()
-    expected = pd.DataFrame(
-        {
-            "id": [1, 2, 3, 4, 5, 6],
-            "date_col": pd.to_datetime(
-                [
-                    "2021-01-01 10:00:00",
-                    None,
-                    None,
-                    None,
-                    None,
-                    "2023-01-01 10:00:00",
-                ]
-            ),
-        }
-    )
-    preview_df = preview_df.sort_values("id").reset_index(drop=True)
-    pd.testing.assert_frame_equal(preview_df, expected)
